Conflicts:
	db/migrate/20110120212424_add_hide_by_default_column_to_ministry_role.rb
	vendor/plugins/mh_common

+++ db/migrate/20110107155626_add_mentor_id_column.rb
@@ -1,9 +1,9 @@
 class AddMentorIdColumn < ActiveRecord::Migration
   def self.up
-    add_column Person.table_name, :person_mentor_id, :integer, :default => 0
+    add_column Person.table_name, :person_mentor_id, :integer, :default => 0 unless Person.columns_hash.has_key?("person_mentor_id")
   end
 
   def self.down
-    remove_column Person.table_name, :person_mentor_id
+    remove_column Person.table_name, :person_mentor_id if Person.columns_hash.has_key?("person_mentor_id")
   end
 end
